Completed cleaning up code/comments.
